refactor: create quantity & breadcrumbs hooks for product slug
